#27 Move tests to `__tests__` folder
