+++ src/__tests__/collect_command.test.js
@@ -1,4 +1,4 @@
-import collectCommand from './collect_command.js'
+import collectCommand from '../collect_command'
 
 it('should do nothing', async () => {
   await expect(() => collectCommand('')).rejects.toThrowError()
